Deleted unused functions and authorization by user type in controllers - 20231229

+++ src/main/java/com/oleg/coupons/controllers/CategoriesController.java
@@ -2,8 +2,12 @@ package com.oleg.coupons.controllers;
 
 import com.oleg.coupons.dto.CategoriesPageResult;
 import com.oleg.coupons.dto.Category;
+import com.oleg.coupons.dto.SuccessfulLoginDetails;
+import com.oleg.coupons.enums.ErrorType;
+import com.oleg.coupons.enums.UserType;
 import com.oleg.coupons.exceptions.ApplicationException;
 import com.oleg.coupons.logic.CategoryLogic;
+import com.oleg.coupons.utils.JWTUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
@@ -21,17 +25,35 @@ public class CategoriesController {
     }
 
     @PostMapping
-    public int addCategory(@RequestBody Category category) throws ApplicationException {
+    public int addCategory(@RequestHeader(value = "Authorization") String token,
+                           @RequestBody Category category) throws Exception {
+        SuccessfulLoginDetails successfulLoginDetails = JWTUtils.decodeJWT(token);
+        UserType userType = successfulLoginDetails.getUserType();
+        if (userType != UserType.ADMIN){
+            throw new ApplicationException(ErrorType.UNAUTHORIZED_REQUEST);
+        }
         return this.categoryLogic.addCategory(category);
     }
 
     @PutMapping
-    public void updateCategory(@RequestBody Category category) throws ApplicationException {
+    public void updateCategory(@RequestHeader(value = "Authorization") String token,
+                               @RequestBody Category category) throws Exception {
+        SuccessfulLoginDetails successfulLoginDetails = JWTUtils.decodeJWT(token);
+        UserType userType = successfulLoginDetails.getUserType();
+        if (userType != UserType.ADMIN){
+            throw new ApplicationException(ErrorType.UNAUTHORIZED_REQUEST);
+        }
         this.categoryLogic.updateCategory(category);
     }
 
     @DeleteMapping("/{id}")
-    public void deleteCategory(@PathVariable("id") int id) throws ApplicationException {
+    public void deleteCategory(@RequestHeader(value = "Authorization") String token,
+                               @PathVariable("id") int id) throws Exception {
+        SuccessfulLoginDetails successfulLoginDetails = JWTUtils.decodeJWT(token);
+        UserType userType = successfulLoginDetails.getUserType();
+        if (userType != UserType.ADMIN){
+            throw new ApplicationException(ErrorType.UNAUTHORIZED_REQUEST);
+        }
         this.categoryLogic.deleteCategory(id);
     }
 
@@ -40,14 +62,15 @@ public class CategoriesController {
         return this.categoryLogic.getAll();
     }
 
-    @GetMapping("/{id}")
-    public Category getCategory(@PathVariable("id") int id) throws ApplicationException {
-        return this.categoryLogic.getById(id);
-    }
-
     @GetMapping("/byFilters")
-    public CategoriesPageResult getByFilters(@RequestParam("page") int page,
-                                             @RequestParam(value = "searchText", required = false) String searchText) throws ApplicationException {
+    public CategoriesPageResult getByFilters(@RequestHeader(value = "Authorization") String token,
+                                             @RequestParam("page") int page,
+                                             @RequestParam(value = "searchText", required = false) String searchText) throws Exception {
+        SuccessfulLoginDetails successfulLoginDetails = JWTUtils.decodeJWT(token);
+        UserType userType = successfulLoginDetails.getUserType();
+        if (userType != UserType.ADMIN){
+            throw new ApplicationException(ErrorType.UNAUTHORIZED_REQUEST);
+        }
         return this.categoryLogic.getByFilters(page, searchText);
     }
 }
