+++ src/main/java/com/oleg/coupons/controllers/UsersController.java
@@ -1,14 +1,16 @@
 package com.oleg.coupons.controllers;
 
-import com.oleg.coupons.dto.User;
+import com.oleg.coupons.dto.SuccessfulLoginDetails;
 import com.oleg.coupons.dto.UserLoginData;
 import com.oleg.coupons.dto.UsersPageResult;
-import com.oleg.coupons.exceptions.ApplicationException;
+import com.oleg.coupons.enums.ErrorType;
+import com.oleg.coupons.enums.UserType;
 import com.oleg.coupons.logic.UserLogic;
+import com.oleg.coupons.utils.JWTUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
-
-import java.util.List;
+import com.oleg.coupons.exceptions.ApplicationException;
+import com.oleg.coupons.dto.User;
 
 @RestController
 @RequestMapping("/users")
@@ -26,40 +28,43 @@ public class UsersController {
         return this.userLogic.addUser(user);
     }
 
-    @PutMapping
-    public void updateUser(@RequestBody User user) throws ApplicationException {
-        this.userLogic.updateUser(user);
-    }
-
-    @DeleteMapping("/{id}")
-    public void deleteUser(@PathVariable("id") int id) throws ApplicationException {
-        this.userLogic.deleteUser(id);
-    }
-
-    @GetMapping
-    public List<User> getAllUsers() throws ApplicationException {
-        return this.userLogic.getAll();
-    }
-
-    @GetMapping("/{id}")
-    public User getUser(@PathVariable("id") int id) throws ApplicationException {
-        return this.userLogic.getById(id);
-    }
-
-    @GetMapping("/byCompanyId")
-    public List<User> getUsersByCompanyId(@RequestParam("companyId") int companyId) throws ApplicationException {
-        return this.userLogic.getByCompanyId(companyId);
-    }
-
     @PostMapping("/login")
     public String login(@RequestBody UserLoginData loginDetails) throws ApplicationException {
         return this.userLogic.login(loginDetails);
     }
 
+    @PutMapping
+    public void updateUser(@RequestHeader("Authorization") String token,
+                           @RequestBody User user) throws Exception {
+        SuccessfulLoginDetails successfulLoginDetails = JWTUtils.decodeJWT(token);
+        UserType userType = successfulLoginDetails.getUserType();
+        if (userType != UserType.ADMIN){
+            throw new ApplicationException(ErrorType.UNAUTHORIZED_REQUEST);
+        }
+        this.userLogic.updateUser(user);
+    }
+
+    @DeleteMapping("/{id}")
+    public void deleteUser(@RequestHeader("Authorization") String token,
+                           @PathVariable("id") int id) throws Exception {
+        SuccessfulLoginDetails successfulLoginDetails = JWTUtils.decodeJWT(token);
+        UserType userType = successfulLoginDetails.getUserType();
+        if (userType != UserType.ADMIN){
+            throw new ApplicationException(ErrorType.UNAUTHORIZED_REQUEST);
+        }
+        this.userLogic.deleteUser(id);
+    }
+
     @GetMapping("/byFilters")
-    public UsersPageResult getByFilters(@RequestParam("page") int page,
+    public UsersPageResult getByFilters(@RequestHeader("Authorization") String token,
+                                        @RequestParam("page") int page,
                                         @RequestParam(value = "companyIds", required = false) Integer[] companyIds,
-                                        @RequestParam(value = "searchText", required = false) String searchText) throws ApplicationException {
+                                        @RequestParam(value = "searchText", required = false) String searchText) throws Exception {
+        SuccessfulLoginDetails successfulLoginDetails = JWTUtils.decodeJWT(token);
+        UserType userType = successfulLoginDetails.getUserType();
+        if (userType != UserType.ADMIN){
+            throw new ApplicationException(ErrorType.UNAUTHORIZED_REQUEST);
+        }
         return this.userLogic.getByFilters(page, companyIds, searchText);
     }
 }
